requirements.txt
openai
python-dotenv
autopep8
speechrecognition
pyaudio
pyttsx3